Generalized Linear Models Fitting with Slab and Shrinkage Estimators

savvy_glm.fit2 is a modified version of glm.fit in the stats package, incorporating custom optimization functions (St_ost, DSh_ost, SR_ost, GSR_ost, and Sh_ost). The user can specify a preferred shrinkage model via the model_class argument. When the user does not explicitly supply a value for model_class, the function defaults to running only "St", "DSh", "SR", and "GSR". If the user explicitly includes "Sh" (e.g. model_class = c("St", "DSh", "SR", "GSR", "Sh")), then Sh_ost is also evaluated.

Usage

savvy_glm.fit2(x, y, weights = rep(1, nobs),
                        model_class = c("St", "DSh", "SR", "GSR", "Sh"),
                        start = NULL, etastart = NULL, mustart = NULL,
                        offset = rep(0, nobs), family = gaussian(),
                        control = list(), intercept = TRUE,
                        use_parallel = TRUE)

Arguments

x

A numeric matrix of predictors. As for glm.fit.

y

A numeric vector of responses. As for glm.fit.

weights

An optional vector of weights to be used in the fitting process. As for glm.fit.

model_class

A character vector specifying the shrinkage model(s) to be used. Allowed values are "St", "DSh", "SR", "GSR", and "Sh". If a single value is provided, only that method is run. If multiple values are provided, the function runs the specified methods in parallel and returns the best one based on AIC. When the user does not explicitly supply a value for model_class, the default is c("St", "DSh", "SR", "GSR") (i.e. "Sh" is not considered).

start

Starting values for the parameters. As for glm.fit.

etastart

Starting values for the linear predictor. As for glm.fit.

mustart

Starting values for the mean. As for glm.fit.

offset

An optional offset to be included in the model. As for glm.fit.

family

A description of the error distribution and link function to be used in the model. As for glm.fit.

control

A list of parameters for controlling the fitting process. As for glm.fit.

intercept

A logical value indicating whether an intercept should be included in the model. As for glm.fit.

use_parallel

Logical. If TRUE, enables parallel execution of the fitting process. Defaults to TRUE. Set to FALSE for serial execution.

Value

The value returned by savvy_glm.fit2 has the same structure as the value returned by glm.fit. It includes the following components:

coefficients

the estimated coefficients.

residuals

the working residuals, that is the residuals in the final iteration of the IWLS fit.

fitted.values

the fitted mean values, obtained by transforming the linear predictors by the inverse of the link function.

R

the upper-triangular factor of the QR decomposition of the weighted model matrix.

rank

the numeric rank of the fitted linear model.

qr

the QR decomposition of the weighted model matrix.

family

the family object used.

linear.predictors

the final linear predictors.

deviance

the deviance of the final model.

aic

the AIC of the final model.

null.deviance

the deviance for the null model.

iter

the number of iterations used.

weights

the final weights used in the fitting process.

prior.weights

the weights initially supplied.

df.residual

the residual degrees of freedom.

df.null

the residual degrees of freedom for the null model.

y

the response vector used.

converged

a logical value indicating whether the IRLS iterations converged.

boundary

a logical value indicating whether the algorithm stopped at a boundary value.

time

the time taken for the fitting process.

chosen_fit

the name of the chosen fitting method based on AIC.

Details

savvy_glm.fit2 extends glm.fit by using custom optimization functions to improve the convergence properties of the iteratively reweighted least squares (IRLS) algorithm. The user may choose to run a specific shrinkage method by supplying a single value to the model_class argument. If multiple values are provided (or if the default is used), the function evaluates a set of custom optimization methods and selects the final model based on the Akaike Information Criterion (AIC). By default, the function considers only St_ost, DSh_ost, SR_ost, and GSR_ost (i.e. model_class = c("St", "DSh", "SR", "GSR")). If the user explicitly includes "Sh" in model_class (for example, model_class = c("St", "DSh", "SR", "GSR", "Sh")), then the method Sh_ost is also evaluated. The fitting process starts with initial parameter values and iterates through the IRLS algorithm. In each iteration, the coefficients are computed using the specified custom methods. The method with the lowest AIC is chosen as the final model, ensuring that the model converges to the best solution given the data and the specified family.

Custom Optimization Methods:

Stein Estimator (St)

Applies a global multiplicative shrinkage factor to all coefficients.

The factor is chosen to reduce MSE based on the overall signal-to-noise ratio.

Simple and fast; works well when all coefficients can be shrunk similarly.

Diagonal Shrinkage (DSh)

Extends the St approach by assigning a separate shrinkage factor to each coefficient.

Each factor is computed from the coefficient size and its associated variance.

More flexible than St, especially when coefficients vary in scale or importance.

Generalized Slab Regression (GSR)

Extends SR by shrinking along multiple directions based on the data.

These directions are usually chosen as leading principal components of the design matrix.

Provides adaptive regularization in settings with collinearity or factor structure.

Shrinkage Estimator (Sh)

Uses a non-diagonal shrinkage matrix derived from solving a Sylvester equation.

Applies shrinkage by transforming the OLS estimator through a matrix that minimizes the MSE.

Included only when "Sh" is specified in the model_class argument.

References

The custom optimization methods used in this function are designed to improve the convergence properties of the GLM fitting process. Marschner, I.C. (2011) glm2: Fitting generalized linear models with convergence problems. The R Journal, Vol. 3/2, pp.12-15.

See also

Author

Ziwei Chen and Vali Asimit
Maintainer: Ziwei Chen <ziwei.chen.3@citystgeorges.ac.uk>