Generalized Linear Models with Slab and Shrinkage Estimators

savvy_glm2 is a modified version of glm2 in the stats package, incorporating custom optimization methods. This function aims to improve convergence properties by selecting the best fitting method based on the Akaike Information Criterion (AIC). In addition, the user can choose to run a specific shrinkage method by providing a value to the model_class argument. When a single value is provided, only that method is executed. When multiple values are given (or when the default is used), the function evaluates the specified methods in parallel and returns the one with the lowest AIC.

Usage

savvy_glm2(formula, family = gaussian, data, weights,
               model_class = c("St", "DSh", "SR", "GSR"), subset,
               na.action, start = NULL, etastart, mustart, offset,
               control = list(...), model = TRUE,
               method = "savvy_glm.fit2", x = FALSE, y = TRUE,
               contrasts = NULL, ...)

Arguments

formula

An object of class "formula": a symbolic description of the model to be fitted. As for glm.

family

A description of the error distribution and link function to be used in the model. As for glm.

data

An optional data frame, list or environment containing the variables in the model. As for glm.

weights

An optional vector of weights to be used in the fitting process. As for glm.

model_class

A character vector specifying the shrinkage model(s) to be used in the underlying fitter savvy_glm.fit2. Allowed values are "St", "DSh", "SR", "GSR", and "Sh". If a single value is provided, only that method is run. If multiple values are provided (or if the default is used), the function evaluates the specified methods in parallel and returns the one with the lowest AIC. When the user does not explicitly supply a value, the default is c("St", "DSh", "SR", "GSR"), and the additional method "Sh" is not considered.

subset

An optional vector specifying a subset of observations to be used in the fitting process. As for glm.

na.action

A function which indicates what should happen when the data contain NAs. As for glm.

start

Starting values for the parameters in the linear predictor. As for glm.

etastart

Starting values for the linear predictor. As for glm.

mustart

Starting values for the vector of means. As for glm.

offset

An optional vector specifying a priori known component to be included in the linear predictor during fitting. As for glm.

control

A list of control parameters to pass to the iterative fitting process. As for glm.

model

A logical value indicating whether the model frame should be included as a component of the returned value. As for glm.

method

The method to be used in fitting the model. The default is savvy_glm.fit2. This method uses iteratively reweighted least squares (IRLS) with custom optimization methods to ensure better convergence by evaluating different fitting methods and selecting the best one based on AIC. As in glm, the alternative method "model.frame" returns the model frame and does no fitting.

x

A logical value indicating whether the model matrix used in the fitting process should be returned as a component of the returned value. As for glm.

y

A logical value indicating whether the response vector used in the fitting process should be returned as a component of the returned value. As for glm.

contrasts

An optional list. See the contrasts.arg of model.matrix.default. As for glm.

...

Additional arguments to be passed to the low level regression fitting functions. As for glm.

savvy_glm2 extends the glm2 function by using custom optimization methods to improve the convergence properties of the iteratively reweighted least squares (IRLS) algorithm. The function evaluates four custom optimization methods: St_ost, DSh_ost, SR_ost, and GSR_ost. The user can also include the additional method Sh_ost by explicitly specifying "Sh" in the model_class argument. The fitting process starts with initial parameter values and iterates through the IRLS algorithm. In each iteration, the coefficients are computed using the specified custom methods. The final model is chosen as the one with the lowest AIC, ensuring that the model converges to the best possible solution given the data and specified family.

Value

The value returned by savvy_glm2 has exactly the same structure as that returned by glm, except for:

method

the name of the fitter function used, which by default is savvy_glm.fit2.

chosen_fit

the name of the chosen fitting method based on AIC.

References

The custom optimization methods used in this function are designed to improve the convergence properties of the GLM fitting process. Marschner, I.C. (2011) glm2: Fitting generalized linear models with convergence problems. The R Journal, Vol. 3/2, pp.12-15. savvy_glm2 uses code from glm2, whose authors are listed in the help documentation for the stats package.

See also